# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SocialCommander API - Social media management for X and Reddit
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [analytics]
      summary: Aggregate analytics across all accounts
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"
...
  /analytics/heatmap:
    get:
      operationId: getEngagementHeatmap
      tags: [analytics]
      summary: Get engagement heatmap by day/hour
      parameters:
        - name: accountId
          in: query
          schema:
            type: ["integer", "null"]
        - name: days
          in: query
          schema:
            type: integer
            default: 90
      responses:
        "200":
          description: Heatmap data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HeatmapCell"
  /analytics/timeseries:
    get:
      operationId: getAnalyticsTimeseries
      tags: [analytics]
      summary: Time series metrics for charting
      parameters:
        - name: accountId
          in: query
          schema:
            type: ["integer", "null"]
        - name: metric
          in: query
          schema:
            type: string
            enum: [impressions, likes, comments, reposts, followers]
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Time series data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeseriesPoint"

  # ── AI ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AnalyticsOverview:
      type: object
      required: [totalImpressions, totalLikes, totalComments, totalReposts, totalPosts, avgEngagementRate, topAccount, growthRate]
      properties:
        totalImpressions:
          type: integer
        totalLikes:
          type: integer
        totalComments:
          type: integer
        totalReposts:
          type: integer
        totalPosts:
          type: integer
        avgEngagementRate:
          type: number
        topAccount:
          type: ["string", "null"]
        growthRate:
          type: number
...
    HeatmapCell:
      type: object
      required: [dayOfWeek, hour, value]
      properties:
        dayOfWeek:
          type: integer
          description: 0=Sunday, 6=Saturday
        hour:
          type: integer
          description: 0-23
        value:
          type: number
    TimeseriesPoint:
      type: object
      required: [date, value]
      properties:
        date:
          type: string
          format: date
        value:
          type: number

    # ── AI SCHEMAS ────────────────────────────────────────────────────────
...